+++ src/SecurityInsights/internal/Update-AzSentinelThreatIntelligenceIndicator.ps1
@@ -129,7 +129,7 @@ THREATINTELLIGENCEPROPERTY <IThreatIntelligenceIndicatorModelForRequestBody>: Th
   [ValidFrom <String>]: Valid from
   [ValidUntil <String>]: Valid until
 .Link
-https://docs.microsoft.com/powershell/module/az.securityinsights/update-azsentinelthreatintelligenceindicator
+https://learn.microsoft.com/powershell/module/az.securityinsights/update-azsentinelthreatintelligenceindicator
 #>
 function Update-AzSentinelThreatIntelligenceIndicator {
 [OutputType([Microsoft.Azure.PowerShell.Cmdlets.SecurityInsights.Models.Api20210901Preview.IThreatIntelligenceInformation])]