+++ src/Resources/Resources/ChangeLog.md
@@ -20,6 +20,10 @@
 
 ## Upcoming Release
 
+## Version 6.5.1
+* Fixed issue introduced in previous fix for `Set-AzPolicySetDefinition` InternalServerError when the initiative is too large [#20238], which will remove space in value.
+* Fixed `Get-AzRoleAssignment` BadRequest when scope is '/' [#20323]
+
 ## Version 6.5.0
 * Added cmdlet `Get-AzADOrganization`
 * Fixed `Set-AzPolicySetDefinition` InternalServerError when the initiative is too large [#20238]